requirements.txt
PyPtt
python-dotenv
pymysql==1.1.1
mysql-connector-python==9.3.0
greenlet==3.2.3 
playwright==1.52.0 
pyee==13.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1